docs(editor): extract composables, lazy-load fflate, and improve UX
Extract useEditorFiles and useFileTreeCrud composables to reduce
component complexity. Lazy-load fflate to remove ~13KB from the shared
docs bundle. Replace backdrop overlay with useClickOutside + Escape
key support. Consolidate infrastructure file constants, deduplicate
generateAppWrapper, alphabetize file tree, and guard activeFile access.
