+++ apps/docs/src/data/editor-defaults.ts
@@ -1,3 +1,11 @@
+// ── Infrastructure file constants ────────────────────────────────────────
+
+/** Files generated by the editor infrastructure (not user-created) */
+export const INFRASTRUCTURE_FILES = new Set(['import-map.json', 'tsconfig.json', 'src/main.ts', 'src/uno.config.ts'])
+
+/** Subset of infrastructure files shown in the "project" folder of the file tree */
+export const REPL_BUILTIN_FILES = ['import-map.json', 'tsconfig.json'] as const
+
 // ── Template files (matching Vuetify Play's v0 template) ────────────────
 
 export function createMainTs (defaultTheme: 'light' | 'dark' = 'light'): string {
